fix(Profit): 수익 금액 출력이 정수로 되도록 수정

+++ src/main/java/blackjack/domain/bet/Profit.java
@@ -10,7 +10,7 @@ public class Profit {
         this.profit = profit;
     }
 
-    public double get() {
+    public int get() {
         return profit;
     }
